# docs/contracts/maekon-web.v1.openapi.yaml
openapi: 3.0.3
info:
  title: Maekon Local Web API
  version: "v1"
  description: |
    Auto-generated from docs/contracts/http-interface-manifest.v1.json.
    Request/response bodies are real schemars-generated DTO schemas
    (full disclosure; the API is loopback-only + token-gated). Cross-crate
    maekon-core field types are rendered as opaque objects (build containment).
  x-maekon-document-version: 10
  x-maekon-updated-at: "2026-07-23"
  x-maekon-routes-file: "crates/maekon-web/src/routes.rs"
  x-maekon-contracts-crate: "crates/maekon-api-contracts"
servers:
  - url: /
paths:
...
  "/api/focus/interruptions":
    get:
      tags:
        - focus
      operationId: focus_get_api_focus_interruptions
      summary: "GET /api/focus/interruptions"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InterruptionDto'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/focus/sessions":
    get:
      tags:
        - focus
      operationId: focus_get_api_focus_sessions
      summary: "GET /api/focus/sessions"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkSessionDto'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/focus/suggestions":
    get:
      tags:
        - focus
      operationId: focus_get_api_focus_suggestions
      summary: "GET /api/focus/suggestions"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocalSuggestionDto'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/privacy/egress-ledger":
    get:
      tags:
        - egress_ledger
      operationId: egress_ledger_get_api_privacy_egress_ledger
      summary: "GET /api/privacy/egress-ledger"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EgressLedgerResponse'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/stats/gui-heatmap":
    get:
      tags:
        - stats
      operationId: stats_get_api_stats_gui_heatmap
      summary: "GET /api/stats/gui-heatmap"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GuiHeatmapCell'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    EgressLedgerEntryDto: {"description":"One egress-ledger row — the audit record of a single egress (or prevented\ncapture) event. Field-for-field the ledger's own nine columns.","properties":{"byte_count":{"description":"Serialized plaintext payload size in bytes. `0` for a `capture_blocked`\nrow (a frame that was deliberately NOT captured — no bytes were sent).","format":"int64","type":"integer"},"consent_state":{"description":"Consent snapshot at the egress moment.","type":"string"},"destination":{"description":"Egress destination / sink target string (e.g. `server.batch_upload`,\n`sync.lan`, or `local.capture` for a prevented capture).","type":"string"},"disposition":{"description":"Disposition: `uploaded`, `blocked`, or `capture_blocked` (a frame\nexcluded before capture — prevented-capture evidence, not egress).","type":"string"},"event_id":{"description":"Associated event id, when known.","type":["string","null"]},"event_type":{"description":"Producing event type (e.g. `Context`, `CrossDeviceSync`, `DeletionEvent`).","type":"string"},"occurred_at":{"description":"Occurrence timestamp (RFC3339).","type":"string"},"recipient_count":{"description":"Number of recipients this egress was delivered to. `0` for a\n`capture_blocked` row (nothing left the device).","format":"int64","type":"integer"},"record_id":{"description":"Caller-generated UUID (`egress_ledger.record_id`, UNIQUE).","type":"string"}},"required":["record_id","event_type","byte_count","recipient_count","destination","disposition","consent_state","occurred_at"],"type":"object"}
...
    EgressLedgerResponse: {"description":"Response body for `GET /api/privacy/egress-ledger`.","properties":{"entries":{"description":"The matched ledger rows. Recent mode: newest-first. Range mode:\noldest-first (matching the underlying reader ordering).","items":{"$ref":"#/components/schemas/EgressLedgerEntryDto"},"type":"array"}},"required":["entries"],"type":"object"}
    ErrorResponse: {"properties":{"code":{"type":"string"},"error":{"type":"string"},"status":{"format":"uint16","maximum":65535,"minimum":0,"type":"integer"}},"required":["code","error","status"],"type":"object"}
...
    GuiHeatmapCell: {"properties":{"count":{"format":"uint32","minimum":0,"type":"integer"},"hour":{"type":"string"}},"required":["hour","count"],"type":"object"}
...
    InterruptionDto: {"properties":{"duration_secs":{"format":"uint64","minimum":0,"type":["integer","null"]},"from_app":{"type":"string"},"from_category":{"type":"string"},"id":{"format":"int64","type":"integer"},"interrupted_at":{"type":"string"},"resumed_at":{"type":["string","null"]},"resumed_to_app":{"type":["string","null"]},"to_app":{"type":"string"},"to_category":{"type":"string"}},"required":["id","interrupted_at","from_app","from_category","to_app","to_category"],"type":"object"}
...
    LocalSuggestionDto: {"properties":{"acted_at":{"type":["string","null"]},"created_at":{"type":"string"},"dismissed_at":{"type":["string","null"]},"id":{"format":"int64","type":"integer"},"payload":true,"shown_at":{"type":["string","null"]},"suggestion_type":{"type":"string"}},"required":["id","suggestion_type","payload","created_at"],"type":"object"}
...
    WorkSessionDto: {"properties":{"category":{"type":"string"},"deep_work_secs":{"format":"uint64","minimum":0,"type":"integer"},"duration_secs":{"format":"uint64","minimum":0,"type":"integer"},"ended_at":{"type":["string","null"]},"id":{"format":"int64","type":"integer"},"interruption_count":{"format":"uint32","minimum":0,"type":"integer"},"primary_app":{"type":"string"},"started_at":{"type":"string"},"state":{"type":"string"}},"required":["id","started_at","primary_app","category","state","interruption_count","deep_work_secs","duration_secs"],"type":"object"}